+++ scaffolds/magic-stripe/template/client/src/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App";
+import "@magiclabs/ui/dist/cjs/index.css";
+import { ThemeProvider } from "@magiclabs/ui";
+
+ReactDOM.render(
+  <React.StrictMode>
+    <ThemeProvider root>
+        <App />
+    </ThemeProvider>
+  </React.StrictMode>,
+  document.getElementById("root")
+);