+++ scaffolds/nuxt/template/layouts/default.vue
@@ -1,0 +1,26 @@
+<template>
+  <div>
+    <Header />
+    <Nuxt />
+  </div>
+</template>
+
+</script>
+
+<style>
+*,
+*::before,
+*::after {
+  box-sizing: border-box;
+  margin: 0;
+}
+* {
+  font-family: sans-serif !important;
+  outline: none;
+}
+.container {
+  max-width: 42rem;
+  margin: 0 auto;
+  padding: 0 10px;
+}
+</style>